refactor: category Controller, ConflictException
