+++ src/test/kotlin/me/insiro/home/server/post/controller/CategoryControllerTest.kt
@@ -1,13 +1,10 @@
 package me.insiro.home.server.post.controller
 
 import me.insiro.home.server.application.domain.Status
-import me.insiro.home.server.application.exception.AbsException
-import me.insiro.home.server.post.dto.category.CategoryDTO
 import me.insiro.home.server.post.dto.category.ModifyCategoryDTO
 import me.insiro.home.server.post.entity.Category
 import me.insiro.home.server.post.entity.JoinedPost
 import me.insiro.home.server.post.entity.Post
-import me.insiro.home.server.post.exception.CategoryNotFoundException
 import me.insiro.home.server.post.service.CategoryService
 import me.insiro.home.server.post.service.PostService
 import me.insiro.home.server.testUtils.AbsControllerTest
@@ -16,14 +13,11 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.mockito.Mockito
 import org.mockito.Mockito.mock
-import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
-import org.springframework.http.ResponseEntity
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import org.springframework.test.web.servlet.setup.MockMvcBuilders
-import org.springframework.web.bind.annotation.*
 
 class CategoryControllerTest : AbsControllerTest("/category") {
     private val postService = mock(PostService::class.java)
@@ -100,49 +94,3 @@ class CategoryControllerTest : AbsControllerTest("/category") {
     }
 }
 
-@RestController
-@RequestMapping("categories")
-class CategoryController(
-    private val categoryService: CategoryService,
-    private val postService: PostService,
-) {
-    @GetMapping
-    fun getAllCategories(): ResponseEntity<List<CategoryDTO>> {
-        val categories = categoryService.findAll().map(::CategoryDTO)
-        return ResponseEntity(categories, HttpStatus.OK)
-    }
-
-    @PostMapping
-    fun createCategory(@RequestBody newCategoryDTO: ModifyCategoryDTO): ResponseEntity<CategoryDTO> {
-        val category = categoryService.create(newCategoryDTO) ?: throw CategoryConflictException(newCategoryDTO.name)
-        return ResponseEntity(CategoryDTO(category), HttpStatus.CREATED)
-    }
-
-    @GetMapping("{name}")
-    fun getCategory(@PathVariable name: String): ResponseEntity<CategoryDTO> {
-        val category = categoryService.findByName(name) ?: throw CategoryNotFoundException(name)
-        return ResponseEntity(CategoryDTO(category), HttpStatus.OK)
-    }
-
-    @PatchMapping("{name}")
-    fun updateCategory(
-        @PathVariable name: String,
-        @RequestBody modifyDTO: ModifyCategoryDTO
-    ): ResponseEntity<CategoryDTO> {
-        val category = categoryService.update(name, modifyDTO) ?: throw CategoryNotFoundException(name)
-        return ResponseEntity(CategoryDTO(category), HttpStatus.OK)
-    }
-
-    @DeleteMapping("{name}")
-    fun deleteCategory(@PathVariable name: String): ResponseEntity<String> {
-        val categoryId = categoryService.delete(name) ?: throw CategoryNotFoundException(name)
-        postService.changeCategory(categoryId, null)
-        return ResponseEntity("success", HttpStatus.OK)
-    }
-
-}
-
-class CategoryConflictException : AbsException {
-    constructor(name: String) : super(HttpStatus.CONFLICT, "Category is Conflicted (name : $name)")
-    constructor(id: Category.Id) : super(HttpStatus.CONFLICT, "Category is Conflicted (id : $id)")
-}
